Moved e2e tests into a new script test called 'ga integration'
