+++ src/integrations/ga/SplitToGa.js
@@ -82,7 +82,7 @@ class SplitToGa {
       const fieldsObject = this.filter(data) && this.mapper(data);
 
       // don't send the hit if it is falsy or invalid when generated by a custom mapper
-      if (!fieldsObject || (this.mapper !== SplitToGa.defaultMapper && !SplitToGa.isValidFieldsObject(fieldsObject)))
+      if (!fieldsObject || (this.mapper !== SplitToGa.defaultMapper && !SplitToGa.validateFieldsObject(fieldsObject)))
         return;
 
       // send the hit
